+++ packages/node/src/indexer/dictionary/v1/algorandDictionaryV1.ts
@@ -1,4 +1,4 @@
-// Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
+// Copyright 2020-2025 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
 import {
@@ -8,7 +8,7 @@ import {
   AlgorandHandler,
   isRuntimeDs,
 } from '@subql/common-algorand';
-import { NodeConfig, DictionaryV1 } from '@subql/node-core';
+import { NodeConfig, DictionaryV1, DsProcessorService } from '@subql/node-core';
 import { AlgorandBlockFilter, AlgorandDataSource } from '@subql/types-algorand';
 import {
   DictionaryQueryCondition,
@@ -17,7 +17,6 @@ import {
 import { sortBy, uniqBy } from 'lodash';
 import { SubqueryProject } from '../../../configure/SubqueryProject';
 import { isBaseHandler, isCustomHandler } from '../../../utils/project';
-import { DsProcessorService } from '../../ds-processor.service';
 
 type GetDsProcessor = DsProcessorService['getDsProcessor'];