+++ packages/node/src/indexer/fetch.module.ts
@@ -1,6 +1,7 @@
-// Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
+// Copyright 2020-2025 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
+import path from 'node:path';
 import { Module } from '@nestjs/common';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import {
@@ -9,110 +10,77 @@ import {
   NodeConfig,
   ConnectionPoolService,
   ConnectionPoolStateManager,
-  IProjectUpgradeService,
   InMemoryCacheService,
   MonitorService,
   CoreModule,
-  IStoreModelProvider,
+  blockDispatcherFactory,
+  DsProcessorService,
+  DynamicDsService,
+  UnfinalizedBlocksService,
+  ProjectService,
+  MultiChainRewindService,
+  DictionaryService,
+  FetchService,
 } from '@subql/node-core';
-import { AlgorandApiConnection, AlgorandApiService } from '../algorand';
-import { SubqueryProject } from '../configure/SubqueryProject';
-import {
-  BlockDispatcherService,
-  WorkerBlockDispatcherService,
-} from './blockDispatcher';
+import { AlgorandApiService } from '../algorand';
+import { BlockchainService } from '../blockchain.service';
 import { AlgorandDictionaryService } from './dictionary';
-import { DsProcessorService } from './ds-processor.service';
-import { DynamicDsService } from './dynamic-ds.service';
-import { FetchService } from './fetch.service';
 import { IndexerManager } from './indexer.manager';
-import { ProjectService } from './project.service';
-import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
 
 @Module({
   imports: [CoreModule],
   providers: [
-    UnfinalizedBlocksService,
     {
-      provide: AlgorandApiService,
+      provide: 'APIService',
       useFactory: AlgorandApiService.init,
       inject: ['ISubqueryProject', ConnectionPoolService, EventEmitter2],
     },
+    {
+      provide: 'IBlockchainService',
+      useClass: BlockchainService,
+    },
+    DsProcessorService,
+    DynamicDsService,
+    {
+      provide: 'IUnfinalizedBlocksService',
+      useClass: UnfinalizedBlocksService,
+    },
+    {
+      useClass: ProjectService,
+      provide: 'IProjectService',
+    },
+    MultiChainRewindService,
     IndexerManager,
     {
       provide: 'IBlockDispatcher',
-      useFactory: (
-        nodeConfig: NodeConfig,
-        eventEmitter: EventEmitter2,
-        projectService: ProjectService,
-        projectUpgradeService: IProjectUpgradeService,
-        apiService: AlgorandApiService,
-        indexerManager: IndexerManager,
-        cacheService: InMemoryCacheService,
-        storeService: StoreService,
-        storeModelProvider: IStoreModelProvider,
-        poiSyncService: PoiSyncService,
-        project: SubqueryProject,
-        dynamicDsService: DynamicDsService,
-        unfinalizedBlocksService: UnfinalizedBlocksService,
-        connectionPoolState: ConnectionPoolStateManager<AlgorandApiConnection>,
-        monitorService?: MonitorService,
-      ) =>
-        nodeConfig.workers
-          ? new WorkerBlockDispatcherService(
-              nodeConfig,
-              eventEmitter,
-              projectService,
-              projectUpgradeService,
-              cacheService,
-              storeService,
-              storeModelProvider,
-              poiSyncService,
-              project,
-              dynamicDsService,
-              unfinalizedBlocksService,
-              connectionPoolState,
-              monitorService,
-            )
-          : new BlockDispatcherService(
-              apiService,
-              nodeConfig,
-              indexerManager,
-              eventEmitter,
-              projectService,
-              projectUpgradeService,
-              storeService,
-              storeModelProvider,
-              poiSyncService,
-              project,
-            ),
+      useFactory: blockDispatcherFactory(
+        path.resolve(__dirname, '../../dist/indexer/worker/worker.js'),
+        [],
+      ),
       inject: [
         NodeConfig,
         EventEmitter2,
         'IProjectService',
         'IProjectUpgradeService',
-        AlgorandApiService,
-        IndexerManager,
         InMemoryCacheService,
         StoreService,
         'IStoreModelProvider',
         PoiSyncService,
         'ISubqueryProject',
         DynamicDsService,
-        UnfinalizedBlocksService,
+        'IUnfinalizedBlocksService',
         ConnectionPoolStateManager,
+        'IBlockchainService',
+        IndexerManager,
+        MultiChainRewindService,
         MonitorService,
       ],
     },
     FetchService,
-    AlgorandDictionaryService,
-    DsProcessorService,
-    DynamicDsService,
     {
-      useClass: ProjectService,
-      provide: 'IProjectService',
+      provide: DictionaryService,
+      useClass: AlgorandDictionaryService,
     },
-    MonitorService,
   ],
 })
 export class FetchModule {}