+++ packages/node/src/subcommands/reindex.module.ts
@@ -1,4 +1,4 @@
-// Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
+// Copyright 2020-2025 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
 import { Module } from '@nestjs/common';
@@ -8,39 +8,55 @@ import {
   ConnectionPoolService,
   ConnectionPoolStateManager,
   DbModule,
+  DsProcessorService,
+  DynamicDsService,
   ForceCleanService,
+  MultiChainRewindService,
   NodeConfig,
   PoiService,
   ReindexService,
   storeModelFactory,
   StoreService,
+  UnfinalizedBlocksService,
 } from '@subql/node-core';
 import { Sequelize } from '@subql/x-sequelize';
 import { AlgorandApiService } from '../algorand';
+import { BlockchainService } from '../blockchain.service';
 import { ConfigureModule } from '../configure/configure.module';
-import { DsProcessorService } from '../indexer/ds-processor.service';
-import { DynamicDsService } from '../indexer/dynamic-ds.service';
 
 @Module({
   providers: [
     {
       provide: 'IStoreModelProvider',
       useFactory: storeModelFactory,
-      inject: [NodeConfig, EventEmitter2, SchedulerRegistry, Sequelize],
+      inject: [NodeConfig, EventEmitter2, Sequelize],
     },
     StoreService,
     ReindexService,
-    ForceCleanService,
-    DynamicDsService,
     PoiService,
+    ForceCleanService,
+    {
+      provide: 'UnfinalizedBlocksService',
+      useClass: UnfinalizedBlocksService,
+    },
+    {
+      provide: 'DynamicDsService',
+      useClass: DynamicDsService,
+    },
     DsProcessorService,
     ConnectionPoolStateManager,
     ConnectionPoolService,
     {
-      provide: AlgorandApiService,
+      // Used to work with DI for unfinalizedBlocksService but not used with reindex
+      provide: 'APIService',
       useFactory: AlgorandApiService.init,
       inject: ['ISubqueryProject', ConnectionPoolService, EventEmitter2],
     },
+    {
+      provide: 'IBlockchainService',
+      useClass: BlockchainService,
+    },
+    MultiChainRewindService,
     SchedulerRegistry,
   ],
   controllers: [],