+++ packages/node/src/subcommands/testing.service.ts
@@ -1,4 +1,4 @@
-// Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
+// Copyright 2020-2025 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
 import { Inject, Injectable } from '@nestjs/common';
@@ -9,12 +9,12 @@ import {
   NestLogger,
   TestRunner,
   IBlock,
+  ProjectService,
 } from '@subql/node-core';
 import { AlgorandDataSource } from '@subql/types-algorand';
-import { AlgorandApi, AlgorandApiService, SafeAPIService } from '../algorand';
+import { AlgorandApi, SafeAPIService } from '../algorand';
 import { SubqueryProject } from '../configure/SubqueryProject';
 import { IndexerManager } from '../indexer/indexer.manager';
-import { ProjectService } from '../indexer/project.service';
 import { BlockContent } from '../indexer/types';
 import { TestingModule } from './testing.module';
 
@@ -53,7 +53,6 @@ export class TestingService extends BaseTestingService<
     await testContext.init();
 
     const projectService: ProjectService = testContext.get('IProjectService');
-    const apiService = testContext.get(AlgorandApiService);
 
     // Initialise async services, we do this here rather than in factories, so we can capture one off events
     await projectService.init();